+++ core/components/minishop3/src/ServiceRegistry.php
@@ -55,6 +55,10 @@ class ServiceRegistry
             'class' => \MiniShop3\Services\ConfigService::class,
             'interface' => null,
         ],
+        'ms3_product_service' => [
+            'class' => \MiniShop3\Services\Product\ProductService::class,
+            'interface' => null,
+        ],
         'ms3_product_data_service' => [
             'class' => \MiniShop3\Services\Product\ProductDataService::class,
             'interface' => null,